Moving Analysis components into their own folder
