+++ src/components/Project/ImageAnalysis/LinePlot.vue
